interactive: extract scope-tree renderer into a substrate-generic backend
Both example binaries carried their own copy of the scope-tree walk
(render_tree/resolve/RItem), differing only in the leaf operators and the
Col/Variable type spellings. Lift the walk into the library as
interactive::backend::render_tree<B>, generic over a Backend trait that
fixes the differential container and supplies the leaf operators
(linear/join/reduce/arrange/inspect/leave_dynamic).

Mirrors DD's container-generic style: generic over C = B::Container with
Time fixed concrete, Col = Collection<'s, Time, C> plain; only the
arrangement type is a (Clone-only) GAT. Region enter/leave stay
container-generic via Enter/Leave<Time, Time, ..Container = C> identity
bounds, so nested scopes recurse with the same backend.

backend::{vec,col} each become a Backend impl plus their substrate leaf
ops; ddir_{vec,col} thin to drivers calling the shared renderer. vec/col
produce identical outputs; reach/tc/scc verified.
