+++ interactive/examples/ddir_col.rs
@@ -1,358 +1,15 @@
-//! DD IR columnar backend: parse, lower, render, execute.
+//! DD IR columnar driver: parse, lower, render (via `interactive::backend::col`), execute.
 
 use mimalloc::MiMalloc;
 
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
 
-mod types {
-    /// A row type backed by Vec but using Strides for columnar bounds.
-    /// This ensures uniform-length rows (common in the IR) get compact
-    /// stride-based offset encoding rather than per-element u64 bounds.
-    #[derive(Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
-    pub struct Row(pub Vec<i64>);
-
-    impl Row {
-        pub fn new() -> Self { Row(Vec::new()) }
-        pub fn push(&mut self, v: i64) { self.0.push(v); }
-    }
-
-    impl std::ops::Deref for Row {
-        type Target = [i64];
-        fn deref(&self) -> &[i64] { &self.0 }
-    }
-
-    impl std::iter::FromIterator<i64> for Row {
-        fn from_iter<I: IntoIterator<Item = i64>>(iter: I) -> Self {
-            Row(iter.into_iter().collect())
-        }
-    }
-
-    impl<'a> IntoIterator for &'a Row {
-        type Item = &'a i64;
-        type IntoIter = std::slice::Iter<'a, i64>;
-        fn into_iter(self) -> Self::IntoIter { self.0.iter() }
-    }
-
-    impl IntoIterator for Row {
-        type Item = i64;
-        type IntoIter = std::vec::IntoIter<i64>;
-        fn into_iter(self) -> Self::IntoIter { self.0.into_iter() }
-    }
-
-    impl columnar::Columnar for Row {
-        type Container = columnar::Vecs<Vec<i64>, columnar::primitive::offsets::Strides>;
-
-        fn into_owned<'a>(other: columnar::Ref<'a, Self>) -> Self {
-            Row(other.into_iter().copied().collect())
-        }
-
-        fn copy_from<'a>(&mut self, other: columnar::Ref<'a, Self>) {
-            self.0.clear();
-            self.0.extend(other.into_iter().copied());
-        }
-    }
-
-    impl interactive::ir::RowLike for Row {
-        fn new() -> Self { Row::new() }
-        fn push(&mut self, v: i64) { Row::push(self, v); }
-        fn as_slice(&self) -> &[i64] { &self.0 }
-        fn extend_from_slice(&mut self, other: &[i64]) { self.0.extend_from_slice(other); }
-    }
-
-    pub type Diff = i64;
-    pub type Time = timely::order::Product<u64, differential_dataflow::dynamic::pointstamp::PointStamp<u64>>;
-}
-use types::*;
+use differential_dataflow::dynamic::pointstamp::PointStamp;
 
 use interactive::parse;
 use interactive::lower;
-
-use differential_dataflow::columnar as columnar_support;
-
-mod columnar {
-    use super::types::*;
-
-    pub use super::columnar_support::*;
-    pub use super::columnar_support::{ValSpine, ValBatcher, ValBuilder};
-
-    pub type DdirUpdate = (Row, Row, Time, Diff);
-    pub type DdirRecordedUpdates = RecordedUpdates<DdirUpdate>;
-
-    pub type ColValSpine<K, V, T, R> = ValSpine<K, V, T, R>;
-    pub type ColValBatcher<K, V, T, R> = ValBatcher<K, V, T, R>;
-    pub type ColValBuilder<K, V, T, R> = ValBuilder<K, V, T, R>;
-    pub type ColValChunker<U> = ValChunker<U>;
-}
-
-mod render {
-    use std::sync::Arc;
-    use timely::order::Product;
-    use timely::dataflow::Scope;
-    use differential_dataflow::Collection;
-    use differential_dataflow::operators::iterate::Variable;
-    use differential_dataflow::dynamic::pointstamp::{PointStamp, PointStampSummary};
-    use differential_dataflow::dynamic::feedback_summary;
-    use differential_dataflow::operators::arrange::{Arranged, TraceAgent};
-    use columnar::Columnar;
-    use super::types::*;
-    use interactive::ir::{LinearOp, RowLike, eval_fields, eval_field_into, eval_condition};
-
-    use super::columnar::{DdirUpdate, DdirRecordedUpdates};
-    use super::columnar::{ColValSpine, ColValBuilder};
-
-    type ConcreteTime = Product<u64, PointStamp<u64>>;
-
-    pub type Col<'scope> = Collection<'scope, ConcreteTime, DdirRecordedUpdates>;
-    type Arr<'scope> = Arranged<'scope, TraceAgent<ColValSpine<Row, Row, ConcreteTime, Diff>>>;
-
-    enum Rendered<'scope> {
-        Collection(Col<'scope>),
-        Arrangement(Arr<'scope>),
-    }
-
-    impl<'scope> Rendered<'scope> {
-        fn collection(&self) -> Col<'scope> {
-            match self {
-                Rendered::Collection(c) => c.clone(),
-                Rendered::Arrangement(a) => {
-                    super::columnar_support::as_recorded_updates::<DdirUpdate>(a.clone())
-                }
-            }
-        }
-        fn arrange(&self) -> Arr<'scope> {
-            match self {
-                Rendered::Arrangement(a) => a.clone(),
-                Rendered::Collection(c) => {
-                    use differential_dataflow::operators::arrange::arrangement::arrange_core;
-                    use super::columnar::{ColValBatcher, ColValChunker};
-                    arrange_core::<_, _, ColValChunker<(Row,Row,Time,Diff)>, ColValBatcher<Row,Row,Time,Diff>, ColValBuilder<Row,Row,Time,Diff>, ColValSpine<Row,Row,Time,Diff>>(c.inner.clone(), timely::dataflow::channels::pact::Pipeline, "Arrange")
-                }
-            }
-        }
-    }
-
-    /// Render a Linear chain: one pass applying the ops in sequence. `level` is
-    /// the op's scope depth — it locates the iteration coord for LiftIter and
-    /// the coordinate position EnterAt's delay lands in.
-    fn render_linear<'scope>(c: Col<'scope>, ops: Vec<LinearOp>, level: usize) -> Col<'scope> {
-        super::columnar::join_function(c, move |k, v, t_in, _d| {
-            use timely::progress::Timestamp;
-            let k: Row = Columnar::into_owned(k);
-            let v: Row = Columnar::into_owned(v);
-            // Materialize input time once so LiftIter can read
-            // the iter coord at the operator's scope depth.
-            let t_owned: Time = Columnar::into_owned(t_in);
-            let iter_at_level: i64 = level
-                .checked_sub(1)
-                .and_then(|idx| t_owned.inner.get(idx).copied())
-                .unwrap_or(0) as i64;
-            let mut results: Vec<(Row, Row, Time, Diff)> = vec![(k, v, Time::minimum(), 1i64)];
-            for op in &ops {
-                let mut next = Vec::new();
-                for (k, v, t, d) in results {
-                    match op {
-                        LinearOp::Project(proj) => {
-                            let i = [k.as_slice(), v.as_slice()];
-                            next.push((eval_fields(&proj.key, &i), eval_fields(&proj.val, &i), t, d));
-                        },
-                        LinearOp::Filter(cond) => {
-                            let i = [k.as_slice(), v.as_slice()];
-                            if eval_condition(cond, &i) { next.push((k, v, t, d)); }
-                        },
-                        LinearOp::Negate => {
-                            next.push((k, v, t, -d));
-                        },
-                        LinearOp::EnterAt(field) => {
-                            let delay = {
-                                let mut r = Row::new();
-                                eval_field_into(field, &[k.as_slice(), v.as_slice()], &mut r);
-                                256 * (64 - (r.as_slice().first().copied().unwrap_or(0) as u64).leading_zeros() as u64)
-                            };
-                            let mut coords = smallvec::SmallVec::<[u64; 1]>::new();
-                            for _ in 0..level.saturating_sub(1) { coords.push(0); }
-                            coords.push(delay);
-                            next.push((k, v, Product::new(0u64, PointStamp::new(coords)), d));
-                        },
-                        LinearOp::LiftIter => {
-                            let mut new_v = v.clone();
-                            new_v.push(iter_at_level);
-                            next.push((k, new_v, t, d));
-                        },
-                    }
-                }
-                results = next;
-            }
-            results.into_iter()
-        })
-    }
-
-    fn render_join<'scope>(l: Arr<'scope>, r: Arr<'scope>, projection: &interactive::parse::Projection) -> Col<'scope> {
-        let proj = projection.clone();
-        use differential_dataflow::operators::join::join_traces;
-        use differential_dataflow::collection::AsCollection;
-        use super::columnar::ValColBuilder;
-        let stream = join_traces::<_, _, _, ValColBuilder<DdirUpdate>>(l, r, move |k, v1, v2, t, d1, d2, c| {
-            use differential_dataflow::difference::Multiply;
-            let d = d1.clone().multiply(d2);
-            let i = [k.as_slice(), v1.as_slice(), v2.as_slice()];
-            let (k2, v2): (Row, Row) = (eval_fields(&proj.key, &i), eval_fields(&proj.val, &i));
-            c.give((k2, v2, t, d));
-        });
-        stream.as_collection()
-    }
-
-    fn render_reduce<'scope>(a: Arr<'scope>, reducer: &interactive::parse::Reducer) -> Arr<'scope> {
-        let reducer = reducer.clone();
-        type ReduceFn = dyn for<'a> Fn(columnar::Ref<'a, Row>, &[(columnar::Ref<'a, Row>, Diff)], &mut Vec<(Row, Diff)>) + Send + Sync;
-        let f: Arc<ReduceFn> = match reducer {
-            interactive::parse::Reducer::Min => Arc::new(|_key, vals, output| {
-                if let Some(min) = vals.iter().map(|(v, _)| v.as_slice()).min() {
-                    output.push((Row(min.to_vec()), 1));
-                }
-            }),
-            interactive::parse::Reducer::Distinct => Arc::new(|_key, _vals, output| { output.push((Row::new(), 1)); }),
-            interactive::parse::Reducer::Count => Arc::new(|_key, vals, output| {
-                let count: Diff = vals.iter().map(|(_, d)| *d).sum();
-                if count != 0 { let mut r = Row::new(); r.push(count); output.push((r, 1)); }
-            }),
-        };
-        a.reduce_abelian::<_, ColValBuilder<_,_,_,_>, ColValSpine<_,_,_,_>, _>(
-            "Reduce",
-            move |k, vals, output| { f(k, vals, output); },
-            |col, key, upds| {
-                use columnar::{Clear, Push};
-                col.keys.clear();
-                col.vals.clear();
-                col.times.clear();
-                col.diffs.clear();
-                for (val, time, diff) in upds.drain(..) { col.push((key, &val, &time, &diff)); }
-                // NOTE: required because push above doesn't group by key, val.
-                *col = std::mem::take(col).consolidate();
-            },
-        )
-    }
-
-    fn render_inspect<'scope>(col: Col<'scope>, label: String) -> Col<'scope> {
-        col.inspect_container(move |event| {
-            if let Ok((_time, container)) = event {
-                for (k, v, t, d) in container.updates.view().iter() {
-                    eprintln!("  [{}] ({:?}, {:?}, {:?}, {:?})", label, <Row as Columnar>::into_owned(k), <Row as Columnar>::into_owned(v), <Time as Columnar>::into_owned(t), <Diff as Columnar>::into_owned(d));
-                }
-            }
-        })
-    }
-
-    // ===== Scope-tree renderer =====
-    //
-    // Walks the scope-tree IR in item order; each `{}` scope is a real timely
-    // region (imports enter it, exports leave it, then the dynamic coordinate
-    // pops). Mirrors ddir_vec's render_tree with columnar types.
-
-    use interactive::scope_ir as st;
-
-    enum RItem<'scope> {
-        Op(Rendered<'scope>),
-        Sub(Vec<Col<'scope>>),
-    }
-
-    fn resolve<'scope>(
-        items: &[RItem<'scope>],
-        imports: &[Col<'scope>],
-        var_cols: &[Col<'scope>],
-        r: &st::Ref,
-    ) -> Rendered<'scope> {
-        match r {
-            st::Ref::Local(i) => match &items[*i] {
-                RItem::Op(Rendered::Collection(c)) => Rendered::Collection(c.clone()),
-                RItem::Op(Rendered::Arrangement(a)) => Rendered::Arrangement(a.clone()),
-                RItem::Sub(_) => panic!("Ref::Local points at a child scope"),
-            },
-            st::Ref::Import(i) => Rendered::Collection(imports[*i].clone()),
-            st::Ref::Var(i) => Rendered::Collection(var_cols[*i].clone()),
-            st::Ref::ChildExport(i, j) => match &items[*i] {
-                RItem::Sub(exports) => Rendered::Collection(exports[*j].clone()),
-                RItem::Op(_) => panic!("Ref::ChildExport points at an operator"),
-            },
-        }
-    }
-
-    pub fn render_tree<'scope>(
-        s: &st::Scope,
-        scope: Scope<'scope, ConcreteTime>,
-        depth: usize,
-        imports: Vec<Col<'scope>>,
-    ) -> Vec<Col<'scope>> {
-        let mut var_handles: Vec<Option<Variable<'scope, ConcreteTime, DdirRecordedUpdates>>> = Vec::new();
-        let mut var_cols: Vec<Col<'scope>> = Vec::new();
-        for _ in &s.vars {
-            let step: Product<u64, PointStampSummary<u64>> = Product::new(0, feedback_summary::<u64>(depth, 1));
-            let (var, col) = Variable::new(scope, step);
-            var_handles.push(Some(var));
-            var_cols.push(col);
-        }
-
-        let mut items: Vec<RItem<'scope>> = Vec::new();
-        for item in &s.items {
-            match item {
-                st::Item::Op(node) => {
-                    let rendered = match node {
-                        st::Node::Linear { input, ops } => {
-                            let c = resolve(&items, &imports, &var_cols, input).collection();
-                            Rendered::Collection(render_linear(c, ops.clone(), depth))
-                        },
-                        st::Node::Concat(refs) => {
-                            let mut c = resolve(&items, &imports, &var_cols, &refs[0]).collection();
-                            for r in &refs[1..] { c = c.concat(resolve(&items, &imports, &var_cols, r).collection()); }
-                            Rendered::Collection(c)
-                        },
-                        st::Node::Arrange(r) => Rendered::Arrangement(resolve(&items, &imports, &var_cols, r).arrange()),
-                        st::Node::Join { left, right, projection } => {
-                            let l = resolve(&items, &imports, &var_cols, left).arrange();
-                            let r = resolve(&items, &imports, &var_cols, right).arrange();
-                            Rendered::Collection(render_join(l, r, projection))
-                        },
-                        st::Node::Reduce { input, reducer } => {
-                            let a = resolve(&items, &imports, &var_cols, input).arrange();
-                            Rendered::Arrangement(render_reduce(a, reducer))
-                        },
-                        st::Node::Inspect { input, label } => {
-                            let c = resolve(&items, &imports, &var_cols, input).collection();
-                            Rendered::Collection(render_inspect(c, label.clone()))
-                        },
-                    };
-                    items.push(RItem::Op(rendered));
-                },
-                st::Item::Sub(child) => {
-                    let child_imports: Vec<Col<'scope>> = child.imports.iter().map(|imp| match &imp.from {
-                        st::Source::Parent(r) => resolve(&items, &imports, &var_cols, r).collection(),
-                        other => panic!("non-root scope with external source {:?}", other),
-                    }).collect();
-                    // Each `{}` scope is a real timely region: imports enter it,
-                    // the child renders inside, exports leave it structurally —
-                    // and then pop the child's dynamic coordinate.
-                    let exported = scope.region_named(&child.name, |region| {
-                        let entered: Vec<_> = child_imports.iter().map(|c| c.clone().enter(region)).collect();
-                        let exports = render_tree(child, region, depth + 1, entered);
-                        exports.into_iter().map(|c| c.leave(scope)).collect::<Vec<_>>()
-                    });
-                    let left: Vec<Col<'scope>> = exported.into_iter().map(|c| super::columnar::leave_dynamic(c, depth + 1)).collect();
-                    items.push(RItem::Sub(left));
-                },
-            }
-        }
-
-        for bind in &s.binds {
-            let c = resolve(&items, &imports, &var_cols, &bind.value).collection();
-            var_handles[bind.var].take().expect("bind: variable already bound").set(c);
-        }
-
-        s.exports.iter().map(|e| resolve(&items, &imports, &var_cols, &e.value).collection()).collect()
-    }
-
-}
-
-use differential_dataflow::dynamic::pointstamp::PointStamp;
+use interactive::backend::col::{render_tree, Row, Diff};
 
 type DdirOuterUpdate = (Row, Row, u64, Diff);
 
@@ -368,7 +25,7 @@ fn run(name: &str, stmts: Vec<parse::Stmt>, n_inputs: usize, nodes: u64, edges: 
     timely::execute_from_args(std::env::args().skip(4), move |worker| {
         use timely::dataflow::InputHandle;
         use timely::container::PushInto;
-        use columnar_support::ValColBuilder;
+        use differential_dataflow::columnar::ValColBuilder;
 
         type OuterBuilder = ValColBuilder<DdirOuterUpdate>;
 
@@ -389,7 +46,7 @@ fn run(name: &str, stmts: Vec<parse::Stmt>, n_inputs: usize, nodes: u64, edges: 
                     interactive::scope_ir::Source::Trace(name) => panic!("ddir_col: Import {:?} not supported in this harness (no trace registry).", name),
                     interactive::scope_ir::Source::Parent(_) => unreachable!("root scope cannot import from a parent"),
                 }).collect();
-                let exports = render::render_tree(&tree.root, inner, 0, root_imports);
+                let exports = render_tree(&tree.root, inner, 0, root_imports);
                 exports[tree_export_idx].clone().leave(scope)
             });
             output.probe_with(&mut probe);
